Remove Kraph and switch to expedia GraphQL library

+++ wallboard/src/main/kotlin/com/sonarsource/dev/quirrus/wallboard/WallboardApp.kt
@@ -311,7 +311,7 @@ fun WallboardApp() {
                                         amountSucceeded,
                                         amountFailed,
                                         totalAmount,
-                                        selectedTasks.first.buildCreatedTimestamp,
+                                        selectedTasks.first.buildCreatedTimestamp.toLong(),
                                         backgroundLoadingInProgress,
                                     )
